requirements.txt
python-dotenv
matplotlib
seaborn
numpy
pandas
requests
pyarrow
fastparquet
yfinance
plotly
Streamlit
boto3
